adjusting insurer4 request

+++ src/main/java/com/referidos/app/segurosref/pojo/bci/QuoteProductBciPojo.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class QuoteProductBciPojo {
 
-    private int IdProducto;
+    private Long IdProducto;
     private String NombreProducto;
     private List<QuoteRateBciPojo> Tarifas;
